refactor(agent): Move in-app agent feature to EE
