+++ web/src/__tests__/server/in-app-agent-stream.servertest.ts
@@ -3,7 +3,7 @@ import { HttpAgent } from "@ag-ui/client";
 import { Agent } from "@mastra/core/agent";
 import { describe, expect, it, vi } from "vitest";
 
-import type { AgUiEvent } from "@/src/features/in-app-agent/schema";
+import type { AgUiEvent } from "@/src/ee/features/in-app-agent/schema";
 
 const adapterEvents = vi.hoisted(() => ({
   items: [] as AgUiEvent[],
@@ -78,7 +78,7 @@ vi.mock("@mastra/mcp", () => ({
   }),
 }));
 
-vi.mock("@/src/features/in-app-agent/server/instrumentation", () => ({
+vi.mock("@/src/ee/features/in-app-agent/server/instrumentation", () => ({
   createInAppAgentInstrumentation:
     instrumentationMocks.createInAppAgentInstrumentation,
 }));
@@ -90,7 +90,7 @@ describe("createAgUiStream", () => {
 
   it("serializes valid events including adapter message snapshots", async () => {
     const { createAgUiStream } =
-      await import("@/src/features/in-app-agent/server/agent");
+      await import("@/src/ee/features/in-app-agent/server/agent");
     const input = {
       threadId: "conversation-1",
       runId: "run-1",
@@ -242,7 +242,7 @@ describe("createAgUiStream", () => {
 
   it("lets HttpAgent subscribers observe streamed run errors", async () => {
     const { createAgUiStream } =
-      await import("@/src/features/in-app-agent/server/agent");
+      await import("@/src/ee/features/in-app-agent/server/agent");
     const input = {
       threadId: "conversation-1",
       runId: "run-1",
