+++ web/src/app/api/in-app-agent/route.ts
@@ -1,4 +1,4 @@
-import inAppAgentHandler from "@/src/features/in-app-agent/server/handler";
+import inAppAgentHandler from "@/src/ee/features/in-app-agent/server/handler";
 
 export const dynamic = "force-dynamic";
 export const runtime = "nodejs";
