+++ web/src/ee/features/in-app-agent/components/InAppAgentMessage.stories.tsx
@@ -1,4 +1,4 @@
-import preview from "../../../../.storybook/preview";
+import preview from "../../../../../.storybook/preview";
 import { InAppAgentMessage } from "./InAppAgentMessage";
 
 const meta = preview.meta({
