+++ web/src/ee/features/in-app-agent/components/InAppAgentWindow.stories.tsx
@@ -1,4 +1,4 @@
-import preview from "../../../../.storybook/preview";
+import preview from "../../../../../.storybook/preview";
 import { useArgs } from "storybook/preview-api";
 import { fn } from "storybook/test";
 import {
