+++ web/src/ee/features/in-app-agent/components/InAppAgentWindow.tsx
